+++ CHANGELOG.md
@@ -10,6 +10,8 @@ CHANGELOG for FlatCAM beta
 26.11.2020
 
 - fixed error in Tools Database when deleting first tool; remade the tool deletion method
+- in Tools Database added tooltips for each column item in the left TreeWidget
+- in Tools Database made sure that editing the Target and Tool diameter in the right side section will update the values in the left TreeWidget
 
 25.11.2020